Move types into GridTypes
