+++ packages/grid/src/Grid.test.tsx
@@ -3,12 +3,12 @@ import TestRenderer from 'react-test-renderer';
 import Grid from './Grid';
 import GridRange from './GridRange';
 import GridRenderer from './GridRenderer';
-import GridTheme, { GridTheme as GridThemeType } from './GridTheme';
+import GridTheme from './GridTheme';
 import GridUtils from './GridUtils';
 import MockGridModel from './MockGridModel';
 import MockGridData from './MockGridData';
-import { VisibleIndex } from './GridMetrics';
 import GridModel from './GridModel';
+import { VisibleIndex, GridTheme as GridThemeType } from './GridTypes';
 
 function makeMockContext(): CanvasRenderingContext2D {
   // Just return a partial mock
