+++ packages/grid/src/GridColorUtils.ts
@@ -1,6 +1,6 @@
 import convert from 'color-convert';
 import { HEX } from 'color-convert/conversions';
-import { GridColor } from './GridTheme';
+import { GridColor } from './GridTypes';
 
 /**
  * Darken the provided colour
