+++ packages/grid/src/ThemeContext.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import { GridTheme as GridThemeType } from './GridTheme';
+import { GridTheme as GridThemeType } from './GridTypes';
 
 export const ThemeContext: React.Context<
   Partial<GridThemeType>
