+++ packages/grid/src/mouse-handlers/GridRowTreeMouseHandler.ts
@@ -3,7 +3,8 @@ import { getOrThrow } from '@deephaven/utils';
 import { EventHandlerResult } from '../EventHandlerResult';
 import Grid from '../Grid';
 import GridMouseHandler from '../GridMouseHandler';
-import GridUtils, { GridPoint } from '../GridUtils';
+import { GridPoint } from '../GridTypes';
+import GridUtils from '../GridUtils';
 
 /**
  * Detect when the tree expand/collapse button is clicked
